+++ src/rpc/mod.rs
@@ -1,6 +1,6 @@
 pub mod api;
-pub mod api_rest;
-pub mod api_solana;
+pub mod api_httpget;
+pub mod api_jsonrpc;
 pub mod server;
 pub mod upstream;
 pub mod workers;