+++ CHANGELOG.md
@@ -14,9 +14,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- bench: add received size ([#29](https://github.com/lamports-dev/alpamayo/pull/29))
-
 ### Breaking
+
+## [0.3.0] - 2025-04-21
+
+### Features
+
+- bench: add received size ([#29](https://github.com/lamports-dev/alpamayo/pull/29))
+- rpc: use jsonrpc feature from richat ([#30](https://github.com/lamports-dev/alpamayo/pull/30))
 
 ## [0.2.1] - 2025-04-17